% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devo_adaptar_query.R
\name{devo_adaptar_query}
\alias{devo_adaptar_query}
\title{Adaptacion de query a formato API}
\usage{
devo_adaptar_query(query)
}
\arguments{
\item{query}{Query a Devo}
}
\description{
Adaptacion de query a formato API
}
\details{
Adaptacion de query a formato API
}
\examples{
devo_adaptar_query("from tabla select campos")
}
