% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devo_dias.R
\name{devo_dias}
\alias{devo_dias}
\title{Peticion a Devo de los ultimos dias}
\usage{
devo_dias(dias, query, bearer)
}
\arguments{
\item{dias}{Dias de consulta desde el tiempo actual}

\item{query}{Query a Devo}

\item{bearer}{Bearer de usuario para conexion por API}
}
\description{
Peticion a Devo de los ultimos dias
}
\details{
Peticion a Devo de los ultimos dias
}
