% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devo_dmy_hms.R
\name{devo_dmy_hms}
\alias{devo_dmy_hms}
\title{Peticion a Devo en formato dmy_hms}
\usage{
devo_dmy_hms(date_from, date_to, query, bearer)
}
\arguments{
\item{date_from}{Fecha inicio en dmy_hms ("01-01-2021 10:00:00")}

\item{date_to}{Fecha fin en dmy_hms ("01-01-2021 15:00:00")}

\item{query}{Query a Devo}

\item{bearer}{Bearer de usuario para conexion por API}
}
\description{
Peticion de datos a Devo, en un intervalo de tiempo definido dmy_hms
}
\details{
Peticion de datos a Devo, en un intervalo de tiempo definido dmy_hms
}
