% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devo_horas.R
\name{devo_horas}
\alias{devo_horas}
\title{Peticion a Devo de las ultimas horas}
\usage{
devo_horas(horas, query, bearer)
}
\arguments{
\item{horas}{Horas de consulta desde el tiempo actual}

\item{query}{Query a Devo}

\item{bearer}{Bearer de usuario para conexion por API}
}
\description{
Peticion de datos a Devo, de las ultimas horas
}
\details{
Peticion de datos a Devo, de las ultimas horas
}
