% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devo_minutos.R
\name{devo_minutos}
\alias{devo_minutos}
\title{Peticion a Devo de los ultimos minutos}
\usage{
devo_minutos(minutos, query, bearer)
}
\arguments{
\item{minutos}{Minutos de consulta desde el tiempo actual}

\item{query}{Query a Devo}

\item{bearer}{Bearer de usuario para conexión por API}
}
\description{
Peticion de datos a Devo, de los ultimos minutos
}
\details{
Peticion de datos a Devo, de los ultimos minutos
}
