% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devo_response.R
\name{devo_response}
\alias{devo_response}
\title{Peticion a Devo en formato epoch}
\usage{
devo_response(date_from, date_to, query, bearer)
}
\arguments{
\item{date_from}{Fecha inicio en epochtime}

\item{date_to}{Fecha fin en epochtime}

\item{query}{Query a Devo}

\item{bearer}{Bearer de usuario para conexión por API}
}
\description{
Peticion de datos a Devo, en un intervalo de tiempo definido en Epoch
}
\details{
Peticion de datos a Devo, en un intervalo de tiempo definido en Epoch
}
