% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devo_ymd_hms.R
\name{devo_ymd_hms}
\alias{devo_ymd_hms}
\title{Peticion a Devo en formato ymd_hms}
\usage{
devo_ymd_hms(date_from, date_to, query, bearer)
}
\arguments{
\item{date_from}{Fecha inicio en ymd_hms ("2021-01-01 10:00:00")}

\item{date_to}{Fecha fin en ymd_hms ("2021-01-01 15:00:00")}

\item{query}{Query a Devo}

\item{bearer}{Bearer de usuario para conexion por API}
}
\description{
Peticion de datos a Devo, en un intervalo de tiempo definido ymd_hms
}
\details{
Peticion de datos a Devo, en un intervalo de tiempo definido ymd_hms
}
